RoadStateIncluded
This is a model which currently stores the road state for each turning
movement on each arm of the road. However I do not use this information
yet, it is merely generated. The next step is to include a timing
structure along with the stage number.
NB. This is essentially still the turning intention high bid model but
it generates more information